+++ src/transformers/css/component.ts
@@ -1,8 +1,8 @@
-import { ValueProperty } from '../types';
+import { Token } from '../types';
 import { Component } from '../../exporters/components/extractor';
 import { formatComponentCodeBlockComment } from '../utils';
 import { ExportableSharedOptions, ExportableTransformerOptions } from '../../types';
-import { getTokenSetTransformer } from '../tokenSetTransformers';
+import { transform } from '../transformer';
 
 /**
  * Map down to a variable object
@@ -16,7 +16,7 @@ export const transformComponentsToCssVariables = (componentName: string, compone
   
   lines.push(`.${componentCssClass} {`)
   const cssVars = components.map((component) => `\t${formatComponentCodeBlockComment(componentName, component, '/**/')}\n${Object.entries(transformComponentTokensToCssVariables(component, options))
-    .map(([variable, value]) => `\t${variable}: ${value.value};`)
+    .map(([name, token]) => `\t${name}: ${token.value};`)
     .join('\n')}`);
   return lines.concat(cssVars).join('\n\n') + '\n}\n';
 };
@@ -26,26 +26,6 @@ export const transformComponentsToCssVariables = (componentName: string, compone
  * @param tokens
  * @returns
  */
-export const transformComponentTokensToCssVariables = (component: Component, options?: ExportableTransformerOptions & ExportableSharedOptions): Record<string, ValueProperty> => {
-  let result = {};
-
-  for (const part in component.parts) {
-    const tokenSets = component.parts[part];
-
-    if (!tokenSets || tokenSets.length === 0) {
-      continue;
-    }
-
-    for (const tokenSet of tokenSets) {
-      const transformer = getTokenSetTransformer(tokenSet);
-
-      if (!transformer) {
-        continue;
-      }
-
-      result = {...result, ...transformer('css', component, part, tokenSet, options)}
-    }
-  }
-  
-  return result;
+export const transformComponentTokensToCssVariables = (component: Component, options?: ExportableTransformerOptions & ExportableSharedOptions): Record<string, Token> => {
+  return transform('css', component, options);
 };