+++ src/transformers/scss/component.ts
@@ -1,9 +1,8 @@
-import { ValueProperty } from '../types';
-import { getSizesFromComponents, getStatesFromComponents, getThemesFromComponents, getTypesFromComponents } from '../css/utils';
+import { Token } from '../types';
+import { getSizesFromComponents, getStatesFromComponents, getThemesFromComponents, getTypesFromComponents, normalizeTokenNameVariableValue } from '../utils';
 import { Component } from '../../exporters/components/extractor';
 import { ExportableSharedOptions, ExportableTransformerOptions } from '../../types';
-import { getTokenSetTransformer } from '../tokenSetTransformers';
-import { normalizeVariableToken } from '../utils';
+import { transform } from '../transformer';
 
 export const transformComponentsToScssTypes = (name: string, components: Component[], options?: ExportableTransformerOptions & ExportableSharedOptions): string => {
   const lines = [];
@@ -23,7 +22,7 @@ export const transformComponentsToScssTypes = (name: string, components: Compone
   // Sizes
   if (sizes && sizes.length > 0) {
     lines.push(
-      `$${name}-sizes: ( ${sizes.map((type) => `"${normalizeVariableToken('size', type, options)}"`).join(', ')} );`
+      `$${name}-sizes: ( ${sizes.map((type) => `"${normalizeTokenNameVariableValue('size', type, options)}"`).join(', ')} );`
     );
   }
 
@@ -44,26 +43,6 @@ export const transformComponentsToScssTypes = (name: string, components: Compone
   return lines.join('\n\n') + '\n';
 }
 
-export const transformComponentTokensToScssVariables = (component: Component, options?: ExportableTransformerOptions & ExportableSharedOptions): Record<string, ValueProperty> => {
-  let result = {};
-
-  for (const part in component.parts) {
-    const tokenSets = component.parts[part];
-
-    if (!tokenSets || tokenSets.length === 0) {
-      continue;
-    }
-
-    for (const tokenSet of tokenSets) {
-      const transformer = getTokenSetTransformer(tokenSet);
-
-      if (!transformer) {
-        continue;
-      }
-
-      result = {...result, ...transformer('scss', component, part, tokenSet, options)}
-    }
-  }
-  
-  return result;
+export const transformComponentTokensToScssVariables = (component: Component, options?: ExportableTransformerOptions & ExportableSharedOptions): Record<string, Token> => {
+  return transform('scss', component, options);
 }