+++ src/types/plugin.d.ts
@@ -1,4 +1,4 @@
-import { CssTransformerOutput } from 'figma-exporter/src/transformers/css';
+import { TransformerOutput } from 'figma-exporter/src/transformers/types';
 import type {  DocumentationObject } from './figma-exporter/src/types';
 //import { HookReturn } from 'figma-exporter/src/transformers/plugin';
 import { webpack } from 'webpack';
@@ -6,8 +6,8 @@ import { webpack } from 'webpack';
 export interface Plugin {
   init: () => void;
   postExtract: (documentationObject: DocumentationObject) => void;
-  postCssTransformer: (documentationObject: DocumentationObject, css: CssTransformerOutput) => void;
-  postScssTransformer: (documentationObject: DocumentationObject, scss: CssTransformerOutput) => void;
+  postCssTransformer: (documentationObject: DocumentationObject, css: TransformerOutput) => void;
+  postScssTransformer: (documentationObject: DocumentationObject, scss: TransformerOutput) => void;
   postIntegration: (documentationObject: DocumentationObject) => HookReturn[] | void;
   modifyWebpackConfig: (webpackConfig: webpack.Configuration) => webpack.Configuration;
   postPreview: (documentationObject: DocumentationObject) => void;